refactor: reorganize webview state management
