+++ src/webview/index.html
@@ -13,7 +13,7 @@
     <script type="importmap" nonce="${nonce}">
       {
         "imports": {
-          "./modules/stateManager.js": "${stateManagerUri}",
+          "./modules/webviewState.js": "${webviewStateModuleUri}",
           "./modules/fileHandlers.js": "${fileHandlersUri}",
           "./modules/uiHandlers.js": "${uiHandlersUri}",
           "./modules/messageHandlers.js": "${messageHandlersUri}",
